ES_recognize_translate_RU/requirements.txt
streamlit
pillow
torch
transformers
requests